+++ tests/test_base.py
@@ -9,11 +9,6 @@ def test_Buzz_require_condition__basic():
         Buzz.require_condition(False, "fail message")
 
 
-def test_Buzz_require_condition__fails_with_explictly_passed_raise_exc_class():
-    with pytest.raises(ValueError, match="You may not pass"):
-        Buzz.require_condition(False, "fail message", raise_exc_class=Exception)
-
-
 def test_Buzz_enforce_defined__basic():
     Buzz.enforce_defined("dummy", "should not fail")
     with pytest.raises(Buzz, match="fail message"):
@@ -28,12 +23,6 @@ def test_Buzz_handle_errors__basic():
     assert "there was a problem" in str(err_info.value)
     assert "intercepted exception" in str(err_info.value)
     assert "ValueError" in str(err_info.value)
-
-
-def test_Buzz_handle_errors__fails_with_explicitly_passed_raise_exc_class():
-    with pytest.raises(ValueError, match="You may not pass"):
-        with Buzz.handle_errors("intercepted exception", raise_exc_class=Exception):
-            pass
 
 
 def test_Buzz_handle_errors__false_re_raise_absorbs_errors():
@@ -56,9 +45,3 @@ def test_Buzz_check_expressions__basic():
     assert "one is not two" in err_msg
     assert "not a problem" not in err_msg
     assert "zero is still zero" in err_msg
-
-
-def test_Buzz_check_exressions__fails_with_explicitly_passed_raise_exc_class():
-    with pytest.raises(ValueError, match="You may not pass"):
-        with Buzz.check_expressions("there will be errors", raise_exc_class=Exception) as check:
-            check(True)